Separete the components in files
