+++ .eslintrc.yml
@@ -1,0 +1,16 @@
+env:
+  browser: true
+  es2021: true
+extends:
+  - 'plugin:react/recommended'
+  - 'airbnb'
+parserOptions:
+  ecmaFeatures:
+    jsx: true
+  ecmaVersion: 12
+  sourceType: module
+plugins:
+  - react
+rules: {
+    'react/jsx-filename-extension': [1, { extensions: ['.js', '.jsx'] }],
+}
