+++ package.json
@@ -8,6 +8,8 @@
     "@testing-library/user-event": "^12.6.0",
     "bootstrap": "^4.5.3",
     "bootstrap-icons": "^1.2.2",
+    "eslint-config-airbnb": "^18.2.1",
+    "jest": "^26.6.3",
     "react": "^17.0.1",
     "react-bootstrap": "^1.4.0",
     "react-dom": "^17.0.1",
@@ -18,7 +20,9 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "eslint:init": "eslint --init",
+    "eslint:fix": "eslint ./ --fix"
   },
   "eslintConfig": {
     "extends": [
@@ -37,5 +41,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "eslint": "^7.19.0",
+    "eslint-plugin-react": "^7.22.0"
   }
 }
